+++ src/coreId/index.js
@@ -1,7 +1,7 @@
 /*
 
 	ard-eventhub
-	by SWR audio lab
+	by SWR Audio Lab
 
 	this file creates a CLI for hashing ids