+++ src/ingest/auth/login/post.js
@@ -1,7 +1,7 @@
 /*
 
 	ard-eventhub
-	by SWR audio lab
+	by SWR Audio Lab
 
 */
 
@@ -27,11 +27,10 @@ module.exports = async (req, res) => {
 		}
 
 		// return ok
+		const expiresIn = parseInt(login.login.expiresIn)
 		return response.ok(req, res, {
-			expiresIn: parseInt(login.login.expiresIn),
-			expires: DateTime.now()
-				.plus({ seconds: parseInt(login.login.expiresIn) })
-				.toISO(),
+			expiresIn,
+			expires: DateTime.now().plus({ seconds: expiresIn }).toISO(),
 
 			token: login.login.idToken,
 			refreshToken: login.login.refreshToken,